backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
langchain>=1.2.0
langchain-groq>=1.1.0
pypdf2==3.0.1
python-dotenv==1.0.1
